requirements.txt
# Основные зависимости для калорийного бота
python-telegram-bot[job-queue]>=20.0
openai>=1.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Минимальные зависимости для CI/CD тестов
pytest>=7.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
bandit>=1.7.0
